Updates to reflect move to github repo.
Updates to zil_01.py, config.py and deploy_contract.py to reflect github repo.
