+++ src/config/config.py
@@ -1,6 +1,6 @@
 #******************************************************************************
 #* Filename:            config.py
-#* Last Updated:        2019-08-30
+#* Last Updated:        2019-09-06
 #* Status:              Prototype
 #* Description:         Config data for zil_01 object
 #* Project:             Zilliqa dApp Project
@@ -13,10 +13,11 @@
 class Config(object):
     #
     #  Default account definitions ...
-    default_account_address = 'b50b61ecb98683083ad80f45bde5d15ea5ab745b'
-    keystore_file = 'data/test_0001_keystore.json'
-    keystore_passphrase = 'test account 0001 passphrase'
+    #  ... update this with your account details.
+    #default_account_address = '123456789abcdef0123456789abcdef012345668'
+    default_account_address = '[Add account address here]'
+    keystore_file = 'data/accounts/test_account_1.json'
+    keystore_passphrase = 'machine economy passphrase 0001'
     #
     #  Default contract defn ...
-    #default_contract_address = '45dca9586598c8af78b191eaa28daf2b0a0b4f43'
     default_contract_address = None
